% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgm.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostics for lsgm objects}
\usage{
diagnostics(fit)
}
\arguments{
\item{fit}{an fitted \code{lsgm} object}
}
\value{
yes
}
\description{
Diagnostic quantities for an \code{lsgm} fit
}
\details{
detials
}
