% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{lsgm-package}
\alias{lsgm-package}
\title{Length selected growth models.}
\description{
Fits growth models to length at age fisheries data, correcting for
gear selectivity, length bin sampling and ageing errors.
}
\author{
Genevieve Phillips
}
