% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgm.R
\name{lsgm}
\alias{lsgm}
\title{Estimation of a Local Structure Graph Model (LSGM)}
\usage{
lsgm(Y, W, X = NULL, burnin = 10000, thin = 100, MCMC = 10000, seed = 4448)
}
\arguments{
\item{Y}{a vector of length N that contains the values of the binary dependent variable.}

\item{W}{an NXN square matrix whose values represent connectivity between pairs of edges}

\item{X}{a NxK data.frame that contains K independent variables for N observations. Must be of class data.frame.}

\item{burnin}{a scalar specifying burnin.}

\item{thin}{a scalar specifying thinning.}

\item{MCMC}{a scalar specifying the number of iterations for the Gibbs sampler.}

\item{seed}{random seed}

\item{par}{a vector of length 2 that contains the starting values for the intercept (b0) and the spatial parameter (eta). Must be numeric.}
}
\value{
the output is a table of estimated coefficients, standard errors and z-statistics, *eta* is the estimated parameter on the spatial term
}
\description{
This function estimates an LSGM, developed in Chyzh and Kaiser (2019)
}
\examples{
data(W)
data(toy_data)
lsgm(Y=as.matrix(toy_data$Y),W=W,X=as.data.frame(toy_data$X))

}
\references{
{
 Chyzh, Olga V. and Mark S. Kaiser. 2019. "A Local Structure Graph Model: Modeling Formation
of Network Edges as a Function of Other Edges." \emph{Political Analysis}. 27 (4): 397--414. \url{https://doi.org/10.1017/pan.2019.8}

Nieman, Mark David, Carla Martinez Machain, Olga V. Chyzh, and Sam Bell. 2021.
"An International Game of Risk: Troop Placement and Major Power Competition." \emph{The Journal of Politics}. \url{https://doi.org/10.1086/711716}
}
}
\keyword{LSGM}
