% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgm.R
\name{lsgmFit}
\alias{lsgmFit}
\title{Growth model for length selected data.}
\usage{
lsgmFit(
  Age,
  Length,
  Aerr,
  Lbreaks,
  fraction = 1,
  selectivity = function(L) 1,
  Aoffset = 0,
  Amin = 0,
  model,
  tmbpars,
  control = list(),
  verbose = FALSE
)
}
\arguments{
\item{Age}{Vector of measured ages}

\item{Length}{Vector of measured lengths}

\item{Aerr}{Ageing error matrix}

\item{Lbreaks}{Internal length bin boundaries}

\item{fraction}{Sampling fraction for each bin}

\item{selectivity}{Selectivity function}

\item{Aoffset}{Ageing offset}

\item{Amin}{Minimum age}

\item{model}{Model indicator}

\item{tmbpars}{Model parameter list}

\item{control}{List of control parameters for nlminb}

\item{verbose}{Enable tracing information}
}
\value{
An object of class \code{lsgm}, with elements
\item{\code{coefficients}}{a named vector of estimated parameters}
\item{\code{cov}}{the covariance of the estimated parameters}
\item{\code{logL}}{log likelihood}
\item{\code{aic}}{Akaike information criteria}
\item{\code{bic}}{Bayesian information criteria}
\item{\code{opt}}{return value from \code{nlminb}}
\item{\code{obj}}{TMB object}
\item{\code{selectivity}}{selectivity function}
\item{\code{model}}{model name}
}
\description{
Fit a growth model to length-at-age fisheries data,
correcting for gear selectivity, length-stratified sampling, and ageing
errors.
}
\seealso{
\code{\link{nlminb}}, \code{\link[TMB]{MakeADFun}}
}
