% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgm.R
\name{lsVB}
\alias{lsVB}
\title{Growth models for length selected data.}
\usage{
lsVB(
  Age,
  Length,
  Aerr,
  Lbreaks,
  fraction = 1,
  selectivity = function(L) 1,
  Aoffset = 0,
  Amin = 0,
  start = list(Linf = 1, K = 1, t0 = 0, CV = 1),
  control = list(),
  verbose = FALSE
)
}
\arguments{
\item{Age}{Vector of measured ages}

\item{Length}{Vector of measured lengths}

\item{Aerr}{Ageing error matrix}

\item{Lbreaks}{Internal length bin boundaries}

\item{fraction}{Sampling fraction for each bin}

\item{selectivity}{Selectivity function}

\item{Aoffset}{Ageing offset}

\item{Amin}{Minimum age}

\item{start}{Initial estimates of model parameters}

\item{control}{List of control parameters for nlminb}

\item{verbose}{Enable tracing information}
}
\value{
An object of class \code{lsVB} with elements
\item{\code{coefficients}}{a named vector of estimated parameters}
\item{\code{cov}}{the covariance of the estimated parameters}
\item{\code{logL}}{log likelihood}
\item{\code{aic}}{Akaike information criteria}
\item{\code{bic}}{Bayesian information criteria}
\item{\code{opt}}{return value from \code{nlminb}}
\item{\code{obj}}{TMB object}
\item{\code{obj}}{selectivity function}
\item{\code{model}}{model name}
\item{\code{call}}{the matched call}
\item{\code{call}}{function to evaluate the growth curve}
\item{\code{call}}{function to evaluate the gradient of the growth curve}
}
\description{
Fit a von Bertalanffy growth model to length at age fisheries
data, correcting for selectivity, length bin sampling and ageing
errors.
}
\details{
Describe here what it does
}
\seealso{
\code{\link{nlminb}}, \code{\link[TMB]{MakeADFun}}
}
