% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgm.R
\name{predict.lsgm}
\alias{predict.lsgm}
\title{Predict method for \code{lsgm} fits}
\usage{
\method{predict}{lsgm}(object, Age, Aoffset = 0)
}
\arguments{
\item{Age}{a vector of (true) ages to predict}

\item{Aoffset}{a vector of age offsets}

\item{fit}{an \code{lsgm} object}
}
\value{
a dataframe with columns
\item{\code{age}}{the age}
\item{\code{length}}{the predicted length}
\item{\code{se}}{the standard error of the predicted length}
\item{\code{sd}}{the standard deviation of a future observation}
}
\description{
Predicted values from a length selected growth model fit
}
\details{
The computes predictions from a fitted length selected growth
model. The supplied ages are assumed to contain no error (but an
offset can be supplied).  The standard error of the mean is
computed by the delta method, and the standard deviation for the
future observations are computed conditional on the estimated CV.
}
