% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgm.R
\name{profile.lsgm}
\alias{profile.lsgm}
\title{Summarizing length selected growth model fits.}
\usage{
\method{profile}{lsgm}(fitted, which, ...)
}
\arguments{
\item{fitted}{an lsgm object}

\item{which}{name or index of a parameter to profile}

\item{...}{additional parameters to pass to \code{\link[TMB]{tmbprofile}}.}
}
\value{
data.frame with parameter and function values.
}
\description{
Calculate the log likelihood profile with respect to a single
parameter for a length selected growth model fit.
}
\details{
This function is a wrapper for \code{\link[TMB]{tmbprofile}}.
}
\seealso{
\code{\link[TMB]{plot.tmbprofile}}, \code{\link[TMB]{confint.tmbprofile}}.
}
