% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsgm.R
\name{summary.lsgm}
\alias{summary.lsgm}
\title{Summarizing length selected growth model fits.}
\usage{
\method{summary}{lsgm}(object, ...)
}
\arguments{
\item{object}{an lsgm object}

\item{...}{currently ignored}
}
\value{
An object of class \code{summary.lsgm}
}
\description{
Summary method for \code{lsgm} objects
}
