% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy spatial data}
\format{
A data frame with 100 rows and 2 variables:
\describe{
  \item{Y}{a binary dependent variable}
  \item{X}{a continuous independent variable}
}
}
\usage{
toy_data
}
\description{
A dataset containing 100 observations of 2 variables
}
\keyword{datasets}
