% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{W}
\alias{W}
\title{Toy spatial connetivity matrix to use with toy_data dataset}
\format{
a 100x100 connectivity matrix
}
\usage{
W
}
\description{
Toy spatial connetivity matrix to use with toy_data dataset
}
\keyword{datasets}
