% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toy_data}
\alias{toy_data}
\title{Toy spatial data}
\format{
A list containing 100 rows and 2 variables:
\describe{
  \item{X}{a real-valued vector of dimension 100x1}
  \item{Y}{a binary-valued vector of dimension 100x1}
  \item{W}{a real-valued matrix of dimension 100x100}
}
}
\usage{
toy_data
}
\description{
A list containing observations for two variables and a connectivity matrix
}
\keyword{datasets}
